+++ backend/Dtos/StatusReadDto.cs
@@ -5,7 +5,7 @@ namespace CommandApi.Dtos
 {
     public partial class StatusReadDto
     {
-        public short Rma { get; set; }
+        public int Rma { get; set; }
         public string Glitch { get; set; }
         public decimal? RepairCost { get; set; }
         public string Status { get; set; }
